+++ Changelog.md
@@ -1,6 +1,8 @@
 # 0.1.7*
 - Add Panic Message to Error Handel
-- Add http.rs to move raw http parseing out of server.rs
+- Add http.rs to move raw http parsing out of server.rs
+- Add optinal Socket Timeout
+- Dont give up on cookie parsing if cookie header is malformed
 
 # 0.1.6
 - Add Example for Logging