+++ Changelog.md
@@ -3,6 +3,9 @@
 - Add http.rs to move raw http parseing out of server.rs
 - Start / Start Threaded returns Option
 - Add .unwrap to all server.starts in examples
+- Add http.rs to move raw http parsing out of server.rs
+- Add optinal Socket Timeout
+- Dont give up on cookie parsing if cookie header is malformed
 
 # 0.1.6
 - Add Example for Logging